djangoapp/requirements.txt
asgiref==3.10.0
Django==5.2.8
djangorestframework==3.16.1
Markdown==3.10
pillow==12.0.0
python-dotenv==1.2.1
sqlparse==0.5.3
psycopg2-binary==2.9.10
openpyxl==3.1.3
dj-stripe>=2.8.0,<3.0.0
stripe>=14.0.0

